backend/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
immutabledict==4.2.1
Markdown==3.7
numpy==2.2.0
ortools==9.11.4210
pandas==2.2.3
pillow==11.0.0
protobuf==5.26.1
psycopg2==2.9.9
PyJWT==2.10.0
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
